// src/app/shared/generated/model/documentCharacteristicCreateUpdate.ts
/**
 * OneCX Document BFF
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DocumentCharacteristicCreateUpdate { 
    id?: string;
    name?: string;
    value?: string;
}