// src/app/shared/generated/model/uploadAttachmentPresignedUrlRequest.ts
/**
 * OneCX Document BFF
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UploadAttachmentPresignedUrlRequest { 
    fileName?: string;
    attachmentId?: string;
}